requirements.txt
opencv-python-headless==4.10.0.84
numpy==1.26.4
pillow==10.4.0
pytesseract==0.3.13
streamlit==1.37.0
tqdm==4.66.4
